refactor(prom): Provide `Parts` to label traits
this commit updates the signature of `mk_stream_labeler()` and
`init_response()` to instead examine the `http::request::Parts` and
`http::response::Parts`, respectively.

this is another step to simplify our label traits.

no implementation of `MkStreamLabel` examines the body of the request.
moreover, none even uses the `req` parameter at all!

the intent of this signature seems reasonable however, we may want to
look at certain requests and opt to ignore them. one could imagine an
`l5d-*` header to refrain from installing telemetry on the response
body, for example.

however, as the last commit showed, we are subject to plenty of pain
related to body generics. let's remove the need for them.

Signed-off-by: katelyn martin <kate@buoyant.io>