refactor(prom): Simplify `NewRecordResponse` marker
`L` and `M` are not contained within `NewRecordResponse`. this
PhantomData marker exists to make this generic over the metric and label
types it is generating.

strictly speaking, there isn't really a `() -> (L, M)` function in play,
though `ExtractParams` is _close_. this is a little bit subjective, but
we can constrain this just as well with a `PhantomData<(L, M)>`.

Signed-off-by: katelyn martin <kate@buoyant.io>