refactor(prom): Make `MkStreamLabel` and `StreamLabel` object-safe
our north star is making our metrics middleware easier to work with.

we want `StreamLabel` to be boxable. this will let us avoid needing to
thread baroque generics throughout our layer / service code, and instead
rely on dynamic dispatch instead of static `L: StreamLabel` bounds.

in order to move in that direction, we must take a small step back first.

these traits must be object-safe in order to do that. for a trait to be
object safe, per the Rust Reference, we may not use any function-level
generics. accordingly, this moves the `<B>` bounds up to the top-level
traits.

this makes some of our type aliases icky, and introduces new generic
parameters to various types/interfaces. we will be able to walk those
back in subsequent commits, now that these traits can be placed in e.g.
`Box<dyn StreamLabel>` values.

NB: as this is an "Zwischenzug"¹ commit, we don't update the app crates.
this commit will fail to compile if the entire workspace is built.

Signed-off-by: katelyn martin <kate@buoyant.io>

1 - https://en.wikipedia.org/wiki/Zwischenzug

Signed-off-by: katelyn martin <kate@buoyant.io>