refactor(prom): `RequestMetrics` and `RequestMetrics` are `L`-generic
this removes a type parameter on `RequestMetrics` and `ResponseMetrics`,
which in practice were being tied to the same `L: StreamLabel`
implementation.

this greatly simplifies code in the `record_response` middleware, and
removes some type aliases from the route- and backend-level metrics
submodules.

see the diff sections below shaped like this:

```diff
-impl<DurL, StatL> RequestMetrics<DurL, StatL>
-where
-    DurL: EncodeLabelSet + Clone + Eq + std::fmt::Debug + std::hash::Hash + Send + Sync + 'static,
-    StatL: EncodeLabelSet + Clone + Eq + std::fmt::Debug + std::hash::Hash + Send + Sync + 'static,
-{
+impl<L: StreamLabel> RequestMetrics<L> {
```

Signed-off-by: katelyn martin <kate@buoyant.io>