refactor(app): Remove backend-level reëxports
this is another slightly subjective change.

i had a fair amount of difficulty learning how the different pieces of
the metrics infrastructure fit together.

a personal roadblock to getting oriented is that many of the label and
metrics structures are reëxported from various places, as are some of
the traits provided by the underlying `linkerd-http-prom` library.

this removes the `pub use` in the backend metrics submodule.

this makes imports in tests mildly more involved, but has the upside
that types live in, and are imported from, a single place.

Signed-off-by: katelyn martin <kate@buoyant.io>