refactor(app): Add `check_..()` fences to route layer
this introduces some calls to the `check_..` family of functions in the
stack library, to help locate where trait-bounds may be failing to be
satisfied.

while this was initially introduced as a debugging change, it made
errors easier to consume by "fencing" the places where a missing
`Service<T>` or `NewService<T>` implementation was.

this resembles some existing calls elsewhere in our codebase, so these
are left in place to help make future changes to our policy routing more
approachable.

Signed-off-by: katelyn martin <kate@buoyant.io>