review(app): Use `dns::Name` in route hostname labels
https://github.com/linkerd/linkerd2-proxy/pull/3258#discussion_r1794135587

this addresses the concerns linked in this thread, and discussed further
in a review conversation.

this refactors the interfaces of `HttpRoute` to make use of the
`linkerd_app_core::dns::Name` type used elsewhere in our networking
stack.

we now accept a `&http::uri::Uri` reference, provided by a request.
then, we inspect the *host* of this request, and retain it for encoding
in our prometheus metrics if it is a valid domain name.

the `Name` type is compatible with SNI, see:
<https://github.com/linkerd/linkerd2-proxy/blob/main/linkerd/dns/name/src/name.rs#L20-L26>

```rust
/// A DNS Name suitable for use in the TLS Server Name Indication (SNI)
/// extension and/or for use as the reference hostname for which to verify a
/// certificate.
///
/// A `Name` is guaranteed to be syntactically valid. The validity rules are
/// specified in [RFC 5280 Section 7.2], except that underscores are also
/// allowed.
```

NB: this is *slightly* different from the approach we'd considered in
https://github.com/linkerd/linkerd2-proxy/pull/3258#discussion_r1794135587,
which would involve calling `NameAddr::from_authority_with_port`,
and subsequently falling back to
`NameAddr::from_authority_with_default_port` if that failed.

this approach is still phrased around the hyper/http url types, but
moves the call to `host()` into our function, sparing us the slightly
spookier invalidated `Option<&str>` argument.

a `#[cfg(test)]`-gated function that accepts an `Option<Name>` is
provided for the sake of test cases, which may want to inspect the
metrics involving a request whose uri includes an authority we would
*not* like to include in our label values.

Co-Authored-By: Oliver Gould <ver@buoyant.io>
Signed-off-by: katelyn martin <kate@buoyant.io>