fix(grpc): Handle timeouts with a DeadlineExceeded error (#3302)
When the proxy encounters gateway timeouts, it reports a gRPC Unavailable status
code. This is at odds with the [grpc
docs](https://grpc.io/docs/guides/deadlines/) which suggest that
DeadlineExceeded should be used.

This change updates the proxy's error handler to return DeadlineExceeded when a
gRPC request encounters a timeout.