feat(client-policy): only allow one opaq route (#3320)
Because the opaq route mocks its selection logic, the first route returned by
the API will always be used. This means that all subsequent routes are
incoherent and cannot be selected.

There's no value in preserving incoherent states in the proxy, so it's
preferable to initially require that the control plane only returns one opaq
route.

This change removes some of the unused matching types and simplifies the opaq
router to only handle one route.