feat(outbound): Configure TLS routes from the policy API (#3341)
Until now, the TLS routing stack was not wire with the API responses. This PR adds mappings from the protobuf API representation to the correct internal types. This enables the proxy to build the correct sidecar type when it receives a TLS protocol response from the discovery.

In addition to that, in a separate commit the TLS rule matching logic has been simplified for our internal types. This can be done because for TLS routes we do not do any real route matching. Furthermore the validation logic in the controller ensures that a TLSRoute has exactly on rule.

Signed-off-by: Zahari Dichev <zaharidichev@gmail.com>