Allow duplicate weighted backend keys
Currently, if the proxy receives two backends with the same metadata,
one of the backends will get dropped because the backend metadata is
used as the key in a hash map. Attempting to then randomly distribute
requests to the backends can panic when selecting the now non-existent
backend.

This is fixed by no longer using backend metadata as a hash map key,
instead generating separate IDs that are stored in a vec to retain the
declared order of backends while also being used to look up the backend
and associated weight independetly.

Validated with new unit tests excercising duplicate backend keys, as well
as a few around the invariants use to store the backends.

Signed-off-by: Scott Fleener <scott@buoyant.io>