Introduce separate structs for distribution
The existing distribution didn't have a great separation between the different kinds, like having some fields that only apply to one of the variants and not the others.

This splits out the distributions into separate structs to make the logic more clear. The structs have been moved to separate files, which makes it much easier to grok the logic for a specific distribution variant.

Signed-off-by: Scott Fleener <scott@buoyant.io>