chore(outbound): set "serviceprofile" route metadata
When synthesizing outbound routes from a service profile, we need to set the
route metadata. We currently use the parent reference.

This change updates these route reference to use a 'serviceprofile' default
route to indicate that it is synthesized. In a followup change, we should
correct this to prefer serviceprofile metadata returned via the API.