refactor(app): `MockBody` convenience functions for test code (#3349)
* refactor(app): add `pending()` helper for mock body

in timeout tests, we construct mock bodies that are backed by a
`Pending<T>` future that will never yield any data.

this introduces a small `pending()` helper to the `MockBody` type used
in test code in the outbound proxy.

Signed-off-by: katelyn martin <kate@buoyant.io>

* refactor(app): add `error()` helper for mock body

in metrics tests, we construct mock bodies that will yield an error
when polled, to examine how telemetry measures certain edge cases.

this introduces a small `error()` helper to the `MockBody` type used
in test code in the outbound proxy.

Signed-off-by: katelyn martin <kate@buoyant.io>

* refactor(app): add `grpc_status()` helper for mock body

in metrics tests, we construct mock bodies that will yield an gRPC
status code in its trailers section
when polled, to examine how telemetry measures gRPC responses.

this refactors the `trailers()` function to more narrowly focus on the
test coverage needed, to help reduce boilerplate in test code.

Signed-off-by: katelyn martin <kate@buoyant.io>

---------

Signed-off-by: katelyn martin <kate@buoyant.io>